# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Шаблон для создания нового сервиса",
        "title": "TEMPLATE",
        "contact": {}
    },
    "host": "localhost:1000",
    "basePath": "/v1",
    "paths": {
...
        "/user/param/get.list": {
            "get": {
                "description": "Метод получения параметров пользователя по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User/Param"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user.User"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/token/get": {
            "get": {
                "description": "Метод получает токен пользователя (аутентификация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User/Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "user.AccessToken": {
            "type": "object",
            "properties": {
                "hours": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "user.RefreshToken": {
            "type": "object",
            "properties": {
                "hours": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "user.Tokens": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/user.AccessToken"
                },
                "refresh": {
                    "$ref": "#/definitions/user.RefreshToken"
                }
            }
        },
        "user.User": {
            "type": "object",
            "required": [
                "company_id",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}